ScriptSimilarityMicroservice/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pydantic~=2.11.3
sentence-transformers~=4.1.0
python-decouple~=3.8
starlette~=0.46.2
transformers~=4.51.3